chore: move to shared folder and address comments
